backend/requirements.txt
fastapi[standard]==0.128.0
deepagents==0.3.9
tavily-python==0.7.21
langchain-ollama==1.0.1
langgraph-checkpoint-sqlite==3.0.3
sqlmodel==0.0.32
